customer manager jpa

+++ src/main/java/org/example/customermanagement/controller/CustomerController.java
@@ -39,7 +39,7 @@ public class CustomerController {
     }
 
     @GetMapping("/{id}/edit")
-    public String update(@PathVariable int id, Model model) {
+    public String update(@PathVariable Long id, Model model) {
         model.addAttribute("customer", customerService.findById(id));
         return "/update";
     }
@@ -51,7 +51,7 @@ public class CustomerController {
     }
 
     @GetMapping("/{id}/delete")
-    public String delete(@PathVariable int id, Model model) {
+    public String delete(@PathVariable Long id, Model model) {
         model.addAttribute("customer", customerService.findById(id));
         return "/delete";
     }
@@ -64,7 +64,7 @@ public class CustomerController {
     }
 
     @GetMapping("/{id}/view")
-    public String view(@PathVariable int id, Model model) {
+    public String view(@PathVariable Long id, Model model) {
         model.addAttribute("customer", customerService.findById(id));
         return "/view";
     }
